Simplify FastAPI app to use service layer
