+++ requirements.txt
@@ -9,3 +9,4 @@ langdetect
 python-telegram-bot
 python-dotenv
 openai
+mcp<1
